+++ .idea/workspace.xml
@@ -17,12 +17,27 @@
     </configurations>
   </component>
   <component name="ChangeListManager">
-    <list default="true" id="d40d1734-8a96-4534-abc9-e0dc7e3a24ab" name="Changes" comment="c bon paul 15h00">
+    <list default="true" id="d40d1734-8a96-4534-abc9-e0dc7e3a24ab" name="Changes" comment="C BON LA PAUL 15h10">
+      <change afterPath="$PROJECT_DIR$/actualisations.c" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/affichages.c" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/bellepou.c" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/coorpixels.c" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/demarrage.c" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/ecrandejeu.c" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/initialisations.c" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/sauvegardes.c" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/tdg.c" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/verifications.c" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/CMakeLists.txt" beforeDir="false" afterPath="$PROJECT_DIR$/CMakeLists.txt" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/cmake-build-debug/.ninja_deps" beforeDir="false" afterPath="$PROJECT_DIR$/cmake-build-debug/.ninja_deps" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/cmake-build-debug/.ninja_log" beforeDir="false" afterPath="$PROJECT_DIR$/cmake-build-debug/.ninja_log" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/cmake-build-debug/CMakeCache.txt" beforeDir="false" afterPath="$PROJECT_DIR$/cmake-build-debug/CMakeCache.txt" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/cmake-build-debug/CMakeFiles/clion-log.txt" beforeDir="false" afterPath="$PROJECT_DIR$/cmake-build-debug/CMakeFiles/clion-log.txt" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/cmake-build-debug/ProjetInfoS3.exe" beforeDir="false" afterPath="$PROJECT_DIR$/cmake-build-debug/ProjetInfoS3.exe" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/cmake-build-debug/Testing/Temporary/LastTest.log" beforeDir="false" afterPath="$PROJECT_DIR$/cmake-build-debug/Testing/Temporary/LastTest.log" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/cmake-build-debug/build.ninja" beforeDir="false" afterPath="$PROJECT_DIR$/cmake-build-debug/build.ninja" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/header.h" beforeDir="false" afterPath="$PROJECT_DIR$/header.h" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/main.c" beforeDir="false" afterPath="$PROJECT_DIR$/main.c" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
@@ -74,6 +89,11 @@
     </key>
   </component>
   <component name="RunManager">
+    <configuration default="true" type="CLionExternalRunConfiguration" factoryName="Application" REDIRECT_INPUT="false" ELEVATE="false" USE_EXTERNAL_CONSOLE="false" PASS_PARENT_ENVS_2="true">
+      <method v="2">
+        <option name="CLION.EXTERNAL.BUILD" enabled="true" />
+      </method>
+    </configuration>
     <configuration name="ProjetInfoS3" type="CMakeRunConfiguration" factoryName="Application" REDIRECT_INPUT="false" ELEVATE="false" USE_EXTERNAL_CONSOLE="false" PASS_PARENT_ENVS_2="true" PROJECT_NAME="ProjetInfoS3" TARGET_NAME="ProjetInfoS3" CONFIG_NAME="Debug" RUN_TARGET_PROJECT_NAME="ProjetInfoS3" RUN_TARGET_NAME="ProjetInfoS3">
       <method v="2">
         <option name="com.jetbrains.cidr.execution.CidrBuildBeforeRunTaskProvider$BuildBeforeRunTask" enabled="true" />
@@ -95,7 +115,7 @@
       <workItem from="1668585033876" duration="15336000" />
       <workItem from="1669382386675" duration="14976000" />
       <workItem from="1669464971622" duration="134000" />
-      <workItem from="1669469271160" duration="2504000" />
+      <workItem from="1669469271160" duration="4390000" />
     </task>
     <task id="LOCAL-00001" summary="AlleZ gros marche">
       <created>1667906483137</created>
@@ -230,7 +250,14 @@
       <option name="project" value="LOCAL" />
       <updated>1669471359304</updated>
     </task>
-    <option name="localTasksCounter" value="20" />
+    <task id="LOCAL-00020" summary="C BON LA PAUL 15h10">
+      <created>1669471835010</created>
+      <option name="number" value="00020" />
+      <option name="presentableId" value="LOCAL-00020" />
+      <option name="project" value="LOCAL" />
+      <updated>1669471835010</updated>
+    </task>
+    <option name="localTasksCounter" value="21" />
     <servers />
   </component>
   <component name="TypeScriptGeneratedFilesManager">
@@ -260,6 +287,7 @@
     <MESSAGE value="commit parce que voila et puis j'ai pas fini" />
     <MESSAGE value="c pas ouf pour paul là" />
     <MESSAGE value="c bon paul 15h00" />
-    <option name="LAST_COMMIT_MESSAGE" value="c bon paul 15h00" />
+    <MESSAGE value="C BON LA PAUL 15h10" />
+    <option name="LAST_COMMIT_MESSAGE" value="C BON LA PAUL 15h10" />
   </component>
 </project>
