+++ cmake-build-debug/.ninja_log
@@ -1,24 +1,19 @@
 # ninja log v5
-1	260	6911672729555200	build.ninja	88cf9e38f0d34ca
-3	191	6910084017571262	CMakeFiles/ProjetInfoS3.dir/source.c.obj	50a92ab30dd422b5
-2	388	6911030058170859	CMakeFiles/ProjetInfoS3.dir/main.c.obj	30056f90c2e5f54e
-389	581	6911030060021623	ProjetInfoS3.exe	3ecfb9d1e06695f4
+1	117	6911769603326019	build.ninja	a6fb69ac2730f4bd
 1	63	6911675916416318	CMakeFiles/ProjetInfoS3.dir/source.c.obj	891c896a67ff7e89
-8	408	6911675919831429	CMakeFiles/ProjetInfoS3.dir/main.c.obj	9cdb0a42ed5174b1
-409	539	6911675921104998	ProjetInfoS3.exe	a2dd1b2e1767132b
-3	428	6911730341418409	CMakeFiles/ProjetInfoS3.dir/main.c.obj	9cdb0a42ed5174b1
-428	606	6911730343167335	ProjetInfoS3.exe	a2dd1b2e1767132b
-4	472	6911731488676400	CMakeFiles/ProjetInfoS3.dir/main.c.obj	9cdb0a42ed5174b1
-473	674	6911731490591395	ProjetInfoS3.exe	a2dd1b2e1767132b
-2	363	6911733363846865	CMakeFiles/ProjetInfoS3.dir/main.c.obj	9cdb0a42ed5174b1
-363	481	6911733364968865	ProjetInfoS3.exe	a2dd1b2e1767132b
-3	344	6911740912131350	CMakeFiles/ProjetInfoS3.dir/main.c.obj	9cdb0a42ed5174b1
-345	457	6911740913202554	ProjetInfoS3.exe	a2dd1b2e1767132b
-4	348	6911743516113189	CMakeFiles/ProjetInfoS3.dir/main.c.obj	9cdb0a42ed5174b1
-349	453	6911743517114059	ProjetInfoS3.exe	a2dd1b2e1767132b
-4	472	6911744133777817	CMakeFiles/ProjetInfoS3.dir/main.c.obj	9cdb0a42ed5174b1
-472	659	6911744135772813	ProjetInfoS3.exe	a2dd1b2e1767132b
-3	558	6911745670114239	CMakeFiles/ProjetInfoS3.dir/main.c.obj	9cdb0a42ed5174b1
-559	686	6911745671384238	ProjetInfoS3.exe	a2dd1b2e1767132b
-2	356	6911749942766633	CMakeFiles/ProjetInfoS3.dir/main.c.obj	9cdb0a42ed5174b1
-356	460	6911749943757050	ProjetInfoS3.exe	a2dd1b2e1767132b
+182	332	6911768964991752	CMakeFiles/ProjetInfoS3.dir/main.c.obj	9cdb0a42ed5174b1
+456	580	6911768967411851	ProjetInfoS3.exe	c0adf4dd1f42fda7
+30	234	6911768963986751	CMakeFiles/ProjetInfoS3.dir/affichages.c.obj	9073d302170bc1a
+23	175	6911768963411761	CMakeFiles/ProjetInfoS3.dir/initialisations.c.obj	ac2c1f356aef5d40
+12	182	6911768963441745	CMakeFiles/ProjetInfoS3.dir/demarrage.c.obj	d3d8fc545fb9485f
+17	195	6911768963512281	CMakeFiles/ProjetInfoS3.dir/coorpixels.c.obj	197fad60f5d496c8
+6	187	6911768963461766	CMakeFiles/ProjetInfoS3.dir/sauvegardes.c.obj	1c338b7dc063c8c
+2	250	6911768964151747	CMakeFiles/ProjetInfoS3.dir/tdg.c.obj	dbf110f9b0190291
+175	333	6911768965006747	CMakeFiles/ProjetInfoS3.dir/actualisations.c.obj	13dc9cc7e53e67d6
+195	347	6911768965151754	CMakeFiles/ProjetInfoS3.dir/ecrandejeu.c.obj	6280cb3679783cc6
+235	365	6911768965331809	CMakeFiles/ProjetInfoS3.dir/bellepou.c.obj	132406a31d596a97
+187	456	6911768966236751	CMakeFiles/ProjetInfoS3.dir/verifications.c.obj	849725aee04c6d4d
+1	114	6911769603326019	build.ninja	a6fb69ac2730f4bd
+2	105	6911769604576070	CMakeFiles/ProjetInfoS3.dir/main.c.obj	9cdb0a42ed5174b1
+5	109	6911769604626011	CMakeFiles/ProjetInfoS3.dir/caavancedoucement.c.obj	6cd29dceb3fe3bb7
+109	228	6911769605761008	ProjetInfoS3.exe	2e90bb3fb67596d8
