+++ cmake-build-debug/CMakeCache.txt
@@ -84,7 +84,7 @@ CMAKE_INSTALL_PREFIX:PATH=C:/Program Files (x86)/ProjetInfoS3
 CMAKE_LINKER:FILEPATH=C:/Users/Paul Rouxel/Desktop/MinGW/bin/ld.exe
 
 //make program
-CMAKE_MAKE_PROGRAM:FILEPATH=C:/Program Files/JetBrains/CLion 2022.2.1/bin/ninja/win/ninja.exe
+CMAKE_MAKE_PROGRAM:FILEPATH=C:/PROGRA~1/JETBRA~1/CLION2~1.1/bin/ninja/win/ninja.exe
 
 //Flags used by the linker during the creation of modules during
 // all build types.
