+++ cmake-build-debug/CMakeFiles/clion-log.txt
@@ -1,10 +1,4 @@
 "C:\Program Files\JetBrains\CLion 2022.2.1\bin\cmake\win\bin\cmake.exe" -DCMAKE_BUILD_TYPE=Debug "-DCMAKE_MAKE_PROGRAM=C:/Program Files/JetBrains/CLion 2022.2.1/bin/ninja/win/ninja.exe" -G Ninja -S "C:\Users\Paul Rouxel\Documents\GitHub\ProjetInfoS3" -B "C:\Users\Paul Rouxel\Documents\GitHub\ProjetInfoS3\cmake-build-debug"
--- The C compiler identification is GNU 8.1.0
--- Detecting C compiler ABI info
--- Detecting C compiler ABI info - done
--- Check for working C compiler: C:/Users/Paul Rouxel/Desktop/MinGW/bin/gcc.exe - skipped
--- Detecting C compile features
--- Detecting C compile features - done
 -- Configuring done
 -- Generating done
 -- Build files have been written to: C:/Users/Paul Rouxel/Documents/GitHub/ProjetInfoS3/cmake-build-debug
