+++ cmake-build-debug/Testing/Temporary/LastTest.log
@@ -1,3 +1,3 @@
-Start testing: Nov 26 15:09 Paris, Madrid
+Start testing: Nov 26 15:42 Paris, Madrid
 ----------------------------------------------------------
-End testing: Nov 26 15:09 Paris, Madrid
+End testing: Nov 26 15:42 Paris, Madrid
